prediction_engine/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
scikit-learn==1.3.1
pandas==2.1.1
numpy==1.26.0
kafka-python==2.0.2
redis==5.0.1
